【修改】修改格式【作者】杨红亮

+++ src/main/java/bss/controller/ob/OBRuleController.java
@@ -65,9 +65,8 @@ public class OBRuleController {
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public String ruleList(@CurrentUser User user, Model model, HttpServletRequest request, @RequestParam(defaultValue = "1") Integer page) {
 		String authType = null;
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				authType = user.getTypeName();
 				// 竞价规则名称
 				String name = (String) request.getParameter("name");
@@ -108,7 +107,6 @@ public class OBRuleController {
 				model.addAttribute("name", name);
 				model.addAttribute("quoteTime", quoteTimeStr);
 				model.addAttribute("intervalWorkday", intervalWorkdayStr);
-			}
 		}
 		model.addAttribute("authType", authType);
 		return "bss/ob/biddingRules/list";
@@ -128,11 +126,9 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public String addRuleUI(@CurrentUser User user, Model model) {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				model.addAttribute("authType", user.getTypeName());
-			}
 		}
 		return "bss/ob/biddingRules/create";
 	}
@@ -154,12 +150,10 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult addRule(OBRule obRule, @CurrentUser User user, Model model) throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				model.addAttribute("authType", user.getTypeName());
 				return service.addRule(obRule, user);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -181,14 +175,12 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult delete(@CurrentUser User user, HttpServletRequest request, Model model) throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				String id = request.getParameter("id");
 				String[] ids = id.split(",");
 				model.addAttribute("authType", user.getTypeName());
 				return service.delete(ids);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -210,13 +202,11 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult setDefaultRule(@CurrentUser User user, HttpServletRequest request, Model model) throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				String id = request.getParameter("id");
 				model.addAttribute("authType", user.getTypeName());
 				return service.updateDefaultRule(id);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -242,9 +232,8 @@ public class OBRuleController {
 			@RequestParam(defaultValue = "1") Integer page)
 			throws ParseException {
 		String authType = null;
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				authType = user.getTypeName();
 				// 设置日期
 				String specialDateStr = request.getParameter("specialDate");
@@ -256,19 +245,18 @@ public class OBRuleController {
 				}
 				// 类型
 				String dateType = request.getParameter("dateType");
-				Map<String, Object> map = new HashMap<String, Object>();
+				Map<String, Object> map = new HashMap<>();
 				map.put("specialDate", specialDate);
 				map.put("dateType", dateType);
 				map.put("page", page);
 				List<OBSpecialDate> list = service.selectAllOBSpecialDate(map);
-				PageInfo<OBSpecialDate> info = new PageInfo<OBSpecialDate>(list);
+				PageInfo<OBSpecialDate> info = new PageInfo<>(list);
 
 				// 将查询的所有信息存放到model域中
 				model.addAttribute("info", info);
 				// 查询数据回显
 				model.addAttribute("specialDateStr", specialDateStr);
 				model.addAttribute("dateType", dateType);
-			}
 		}
 		model.addAttribute("authType", authType);
 		return "bss/ob/biddingRules/holiday";
@@ -289,11 +277,9 @@ public class OBRuleController {
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public String createSpecialdateUI(@CurrentUser User user, Model model) {
 		String authType = null;
-		if (user != null) {
+		if (user != null&& "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				authType = user.getTypeName();
-			}
 		}
 		model.addAttribute("authType", authType);
 		return "bss/ob/biddingRules/createSpecialdate";
@@ -316,11 +302,9 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult addSpecialdate(@CurrentUser User user,OBSpecialDate obSpecialDate, HttpServletRequest request) throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				return service.addSpecialdate(obSpecialDate, request, user);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -341,13 +325,11 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult deleteSpecialDate(@CurrentUser User user, HttpServletRequest request) throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				String id = request.getParameter("id");
 				String[] ids = id.split(",");
 				return service.deleteSpecialDate(ids);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -368,16 +350,14 @@ public class OBRuleController {
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public String editobRule(@CurrentUser User user, Model model, String id) {
 		String authType = null;
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				authType = user.getTypeName();
 				if (StringUtils.isEmpty(id)) {
 				} else {
 					OBRule obRule = service.editObRule(id);
 					model.addAttribute("obRule", obRule);
 				}
-			}
 		}
 		model.addAttribute("authType", authType);
 		return "bss/ob/biddingRules/editRule";
@@ -400,11 +380,9 @@ public class OBRuleController {
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult updateobRule(@CurrentUser User user, OBRule obRule)
 			throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				return service.updateobRule(obRule);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -425,15 +403,13 @@ public class OBRuleController {
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public String editSpecialdate(@CurrentUser User user, Model model, String id) {
 		String authType = null;
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				authType = user.getTypeName();
 				if (StringUtils.isEmpty(id)) {
 				} else {
 					OBSpecialDate specialDate = service.editSpecialdate(id);
 					model.addAttribute("specialDate", specialDate);
-				}
 			}
 		}
 		model.addAttribute("authType", authType);
@@ -457,11 +433,9 @@ public class OBRuleController {
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult updateSpecialdate(@CurrentUser User user,
 			OBSpecialDate obSpecialDate) throws Exception {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				return service.updateobSpecialDate(obSpecialDate);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
@@ -481,11 +455,9 @@ public class OBRuleController {
 	@SystemControllerLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	@SystemServiceLog(description = StaticVariables.OB_PROJECT_NAME, operType = StaticVariables.OB_PROJECT_NAME_SIGN)
 	public JdcgResult checkNameUnique(@CurrentUser User user, String name) {
-		if (user != null) {
+		if (user != null && "4".equals(user.getTypeName())) {
 			// 判断是否 是资源服务中心
-			if ("4".equals(user.getTypeName())) {
 				return service.checkNameUnique(name);
-			}
 		}
 		return JdcgResult.build(500, StaticVariables.OB_PROJECT_SHOW);
 	}
